[Update] Users Contact [Create] Mailer

+++ app/controllers/contacts_controller.rb
@@ -18,29 +18,41 @@ class ContactsController < ApplicationController
 	end
 
 	def index
-		@contact = Contact.new #あとで削除
-		@users = User.page(params[:page]).reverse_order
-		@contacts = Contact.page(params[:page]).reverse_order
+		if current_user.admin?#仮if文
+			@users = User.page(params[:page]).reverse_order
+			@contacts = Contact.page(params[:page]).reverse_order
+		else
+			redirect_to '/products'
+		end
 	end
 
 	def show
-		@user = User.find(params[:user_id])
-		@contact = Contact.find(params[:id])
+		if current_user.admin?#仮if文
+			@user = User.find(params[:user_id])
+			@contact = Contact.find(params[:id])
+		else
+			redirect_to '/products'
+		end
 	end
 # 問い合わせ返信
 	def update
 		user = User.find(params[:user_id])
-		contact = Contact.find(params[:id])
-		if contact.update(contact_params)
+		@contact = Contact.find(params[:id])
+		if @contact.update(contact_params)
+			ContactMailer.contact_mail(@contact).deliver#Mailer呼び出し
 			flash[:success] = '返信しました'
 			redirect_to "/contacts"
 		end
 	end
 
 	def destroy
-	    contact = Contact.find(params[:id])
-	    contact.destroy
-	    redirect_to '/contacts'
+		if current_user.admin?#仮if文
+		    contact = Contact.find(params[:id])
+		    contact.destroy
+		    redirect_to '/contacts'
+		else
+			redirect_to '/products'
+		end
   	end
